requirements.txt
mediapipe==0.10.3
numpy==1.23.5
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
Pillow==9.5.0
setuptools==58.1.0
tensorflow==2.17.0
keras==3.4.1
websockets==11.0.3
